break up Loggers.scala
