+++ main/util/src/mill/util/ColorLogger.scala
@@ -1,0 +1,10 @@
+package mill.util
+
+import mill.api.Logger
+
+
+trait ColorLogger extends Logger {
+  def infoColor: fansi.Attrs
+  def errorColor: fansi.Attrs
+}
+
